+++ packages/core/src/hooks/useRegistryData/examples/DummyComponent.tsx
@@ -1,10 +1,4 @@
-import type {
-  FloorRegistryEntry,
-  AreaRegistryEntry,
-  DeviceRegistryEntry,
-  EntityRegistryEntry,
-  EntityRegistryDisplayEntry,
-} from "@hakit/core";
+import type { FloorRegistryEntry, AreaRegistryEntry, DeviceRegistryEntry, EntityRegistryDisplayEntry } from "@hakit/core";
 import type { HassServices, HassService } from "home-assistant-js-websocket";
 
 export function DummyComponentReturn(props: {
@@ -12,7 +6,6 @@ export function DummyComponentReturn(props: {
   areas: Record<string, AreaRegistryEntry>;
   devices: DeviceRegistryEntry;
   services: HassServices;
-  entitiesRegistry: EntityRegistryEntry;
   entitiesRegistryDisplay: EntityRegistryDisplayEntry;
 }) {
   return props;
@@ -31,10 +24,6 @@ export function DummyComponentDevices(props: Partial<DeviceRegistryEntry>) {
 }
 
 export function DummyComponentServices(props: Partial<HassService>) {
-  return props;
-}
-
-export function DummyComponentEntitiesRegistry(props: Partial<EntityRegistryEntry>) {
   return props;
 }