+++ packages/core/src/hooks/useRegistryData/index.ts
@@ -3,7 +3,7 @@ import { useStore, Store } from "@core";
 /**
  * The allowable registry data keys that can be retrieved via the hook.
  */
-export type RegistryDataKey = "entitiesRegistry" | "entitiesRegistryDisplay" | "services" | "areas" | "devices" | "floors";
+export type RegistryDataKey = "entitiesRegistryDisplay" | "services" | "areas" | "devices" | "floors";
 
 /**
  * React hook to access registry-like data slices from the internal store with proper typing.